requirements.txt
tweepy==3.9.0
Flask==1.1.2
psycopg2-binary==2.8.5